venv_requirements.txt
absl-py==1.1.0
aiohttp==3.8.3
aiosignal==1.3.1
apache-beam==2.40.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
argostranslate
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
audioread==2.1.9
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cloudpickle==2.1.0
colorama==0.4.5
commonmark==0.9.1
comtypes==1.1.14
contextlib2==21.6.0
crcmod==1.7
cycler==0.11.0
Cython==0.29.30
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.1.1
dm-tree==0.1.7
docopt==0.6.2
einops==0.6.0
elevenlabslib
entrypoints==0.4
et-xmlfile==1.1.0
etils==0.6.0
fastavro==1.5.2
fastjsonschema==2.16.1
ffmpeg-python==0.2.0
filelock==3.9.0
flatbuffers==1.12
fonttools==4.33.3
frozenlist==1.3.3
future==0.18.2
gast==0.4.0
gin-config==0.5.0
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.3
greenlet==2.0.1
grpcio==1.47.0
h5py==3.7.0
hdfs==2.7.0
httplib2==0.20.4
huggingface-hub==0.11.1
ibm-cloud-sdk-core==3.16.1
ibm-watson==6.1.0
idna==3.3
imbalanced-learn==0.10.0
imblearn==0.0
importlib-metadata==4.12.0
importlib-resources==5.8.0
inflect==5.3.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.9.1
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kaggle==1.5.12
keras==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
langid==1.1.6
libclang==14.0.1
librosa==0.8.1
llvmlite==0.39.0
lvis==0.5.3
lxml==4.9.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
more-itertools==9.0.0
multidict==6.0.4
multiprocess==0.70.9
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.6.3
notebook==6.4.12
numba==0.56.0
numpy
oauth2client==4.1.3
oauthlib==3.2.0
openai==0.27.4
opencv-python==4.5.5.64
opencv-python-headless==4.7.0.68
openpyxl==3.0.10
opt-einsum==3.3.0
orjson==3.7.5
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
pkgutil_resolve_name==1.3.10
playwright==1.29.1
plotly==5.14.0
pooch==1.6.0
portalocker==2.4.0
progressbar==2.5
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.30
proto-plus==1.20.6
protobuf==3.19.4
psutil==5.9.1
py-cpuinfo==8.0.0
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
pycocotools==2.0.4
pycparser==2.21
pydantic==1.10.4
pydot==1.4.2
pydub==0.25.1
pyee==9.0.4
Pygments==2.12.0
PyJWT==2.6.0
pymongo==3.12.3
pynndescent==0.5.7
pyparsing==2.4.7
pypiwin32==223
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyreadline3==3.4.1
pyrsistent==0.18.1
python-dateutil==2.8.2
python-slugify==6.1.2
pyttsx3==2.90
pytz==2022.1
pywin32==304
pywinpty==2.0.7
PyYAML==5.4.1
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
regex==2022.6.2
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.3.1
rich==13.0.1
rotary-embedding-torch==0.2.1
rsa==4.8
sacrebleu==2.1.0
semantic-version==2.10.0
Send2Trash==1.8.0
sentencepiece==0.1.96
seqeval==1.2.2
setuptools-rust==1.5.2
simpleaudio==1.0.4
six==1.16.0
sounddevice==0.4.3
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
tabulate==0.8.10
tenacity==8.2.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-addons==0.17.1
tensorflow-datasets==4.6.0
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.1
tensorflow-hub==0.12.0
tensorflow-io==0.26.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.9.0
tensorflow-model-optimization==0.7.2
tensorflow-text==2.9.0
termcolor==1.1.0
terminado==0.15.0
text-unidecode==1.3
tf-models-official==2.9.2
tf-slim==1.1.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.13.2
toml==0.10.2
tornado==6.2
tqdm==4.62.3
traitlets==5.3.0
transformers==4.25.1
typeguard==2.13.3
typing_extensions==4.2.0
umap-learn==0.5.2
Unidecode==1.3.2
uritemplate==4.1.1
urllib3==1.26.7
visdom==0.1.8.9
wcwidth==0.2.5
webencodings==0.5.1
webrtcvad==2.0.10
websocket-client==1.1.0
Werkzeug==2.1.2
whisper==1.0.0
widgetsnbextension==3.6.1
wrapt==1.14.1
yarl==1.8.2
zipp==3.8.0
pvporcupine==2.1.4
whisper-openai==1.0.0
textblob==0.17.1
langid==1.1.6
SpeechRecognition==3.10.0
pandas
pyaudio==0.2.13
pygame==2.3.0
pyttsx3 
translators
openai
ibm-cloud-sdk-core
ibm-watson
librosa
fsspec
coqpit
trainer
pysbd
matplotlib>=3.7.1
scipy >= 1.10.1
scikit-learn >= 1.2.2
plotly 
python-dotenv
langchain
newsapi-python
wikipedia
geocoder